improve folder structure and ui for add comment and comment list
