--- src/pages/Task/components/MainContent/Activity/index.ts
@@ -1,1 +1,0 @@
-export { default as Activity } from './Activity';
