+++ src/pages/Task/components/MainContent/Comments/AddComment/index.ts
@@ -1,0 +1,1 @@
+export { default as AddComment } from './AddComment';
