+++ src/pages/Task/components/MainContent/MainContent.tsx
@@ -1,10 +1,18 @@
 import React from 'react';
+import { ListRounded as ListRoundedIcon } from '@material-ui/icons';
 
-import { AddDescription } from './AddDescription';
-import { Activity } from './Activity';
-import { StyledContainer } from './MainContent.styles';
 import { Comments } from './Comments';
+import { AddDescription } from './AddDescription';
 import { IComment } from '../../utils';
+
+import {
+  StyledContainer,
+  StyledTitleContainer,
+  StyledTitle,
+  StyledButton,
+  StyledIcon,
+  StyledActivityContainer,
+} from './MainContent.styles';
 
 interface IProps {
   comments: IComment[];
@@ -14,7 +22,13 @@ const MainContent: React.FC<IProps> = ({ comments }) => {
   return (
     <StyledContainer>
       <AddDescription />
-      <Activity />
+      <StyledActivityContainer>
+        <StyledTitleContainer>
+          <StyledIcon component={<ListRoundedIcon />} />
+          <StyledTitle>Activity</StyledTitle>
+          <StyledButton>Show details</StyledButton>
+        </StyledTitleContainer>
+      </StyledActivityContainer>
       <Comments comments={comments} />
     </StyledContainer>
   );
